requirements.txt
flask==3.0.2
sentence-transformers==2.5.1
qdrant-client==1.7.3
openai==1.12.0
PyMuPDF==1.23.26
python-dotenv==1.0.1
reportlab==4.1.0

